+++ api/api-booking/src/main/java/com/pgms/apibooking/domain/payment/dto/request/ConfirmVirtualIncomeRequest.java
@@ -1,0 +1,8 @@
+package com.pgms.apibooking.domain.payment.dto.request;
+
+public record ConfirmVirtualIncomeRequest (
+	String createdAt,
+	String status,
+	String orderId
+){
+}
