+++ api/api-booking/src/main/java/com/pgms/apibooking/domain/payment/dto/request/RefundAccountRequest.java
@@ -1,0 +1,16 @@
+package com.pgms.apibooking.domain.payment.dto.request;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record RefundAccountRequest(
+
+	@NotBlank(message = "[은행] 선택은 필수입니다.")
+	String bank,
+
+	@NotBlank(message = "[환불 계좌] 입력은 필수입니다.")
+	String accountNumber,
+
+	@NotBlank(message = "[예금주] 입력은 필수입니다.")
+	String holderName
+) {
+}
