+++ api/api-booking/src/main/java/com/pgms/apibooking/domain/payment/dto/response/PaymentCancelDetailResponse.java
@@ -1,0 +1,8 @@
+package com.pgms.apibooking.domain.payment.dto.response;
+
+public record PaymentCancelDetailResponse(
+	String cancelReason,
+	int cancelAmount,
+	String canceledAt
+) {
+}
