+++ api/api-booking/src/main/java/com/pgms/apibooking/domain/payment/dto/response/PaymentCardResponse.java
@@ -1,0 +1,8 @@
+package com.pgms.apibooking.domain.payment.dto.response;
+
+public record PaymentCardResponse(
+	String number,
+	int installmentPlanMonths,
+	boolean isInterestFree
+) {
+}
