+++ api/api-booking/src/main/java/com/pgms/apibooking/domain/payment/dto/response/PaymentFailResponse.java
@@ -1,0 +1,8 @@
+package com.pgms.apibooking.domain.payment.dto.response;
+
+public record PaymentFailResponse(
+	String errorCode,
+	String errorMessage,
+	String orderId
+) {
+}
