+++ api/api-booking/src/main/java/com/pgms/apibooking/domain/payment/dto/response/PaymentVirtualResponse.java
@@ -1,0 +1,10 @@
+package com.pgms.apibooking.domain.payment.dto.response;
+
+public record PaymentVirtualResponse(
+	// 토스 응답이라 이름변경x
+	String accountNumber,
+	String bankCode,
+	String customerName,
+	String dueDate
+) {
+}
