+++ api/api-booking/src/main/java/com/pgms/apibooking/domain/payment/dto/response/RefundAccountResponse.java
@@ -1,0 +1,8 @@
+package com.pgms.apibooking.domain.payment.dto.response;
+
+public record RefundAccountResponse(
+	String bankCode,
+	String accountNumber,
+	String holderName
+) {
+}
