+++ api/api-booking/src/main/java/com/pgms/apibooking/domain/seat/dto/response/AreaResponse.java
@@ -1,0 +1,22 @@
+package com.pgms.apibooking.domain.seat.dto.response;
+
+import com.pgms.coredomain.domain.event.EventSeat;
+
+import java.util.List;
+
+public record AreaResponse(
+	String type,
+	int price,
+//	int availableSeatCount,
+	List<SeatResponse> seats
+) {
+
+	public static AreaResponse from(List<EventSeat> seats) {
+		return new AreaResponse(
+			seats.get(0).getEventSeatArea().getSeatAreaType().getDescription(),
+			seats.get(0).getEventSeatArea().getPrice(),
+//			seats.get(0).getEventSeatArea().getAvailableSeatCount(),
+			seats.stream().map(SeatResponse::from).toList()
+		);
+	}
+}
