+++ api/api-booking/src/main/java/com/pgms/apibooking/domain/seat/dto/response/SeatResponse.java
@@ -1,0 +1,16 @@
+package com.pgms.apibooking.domain.seat.dto.response;
+
+import com.pgms.coredomain.domain.event.EventSeat;
+
+public record SeatResponse(
+	String name,
+	boolean isAvailable
+) {
+
+	public static SeatResponse from(EventSeat seat) {
+		return new SeatResponse(
+			seat.getName(),
+			seat.isAvailable()
+		);
+	}
+}
