--- api/api-booking/src/main/java/com/pgms/apibooking/dto/request/ConfirmVirtualIncomeRequest.java
@@ -1,8 +1,0 @@
-package com.pgms.apibooking.dto.request;
-
-public record ConfirmVirtualIncomeRequest (
-	String createdAt,
-	String status,
-	String orderId
-){
-}
