--- api/api-booking/src/main/java/com/pgms/apibooking/dto/request/PaymentCancelRequest.java
@@ -1,14 +1,0 @@
-package com.pgms.apibooking.dto.request;
-
-import java.util.Optional;
-
-public record PaymentCancelRequest(
-	String cancelReason,
-	int cancelAmount,
-	Optional<RefundAccountRequest> refundReceiveAccount
-) {
-
-	public static PaymentCancelRequest of(String cancelReason, int cancelAmount, Optional<RefundAccountRequest> refundReceiveAccount) {
-		return new PaymentCancelRequest(cancelReason, cancelAmount, refundReceiveAccount);
-	}
-}
