--- api/api-booking/src/main/java/com/pgms/apibooking/dto/request/SeatsGetRequest.java
@@ -1,6 +1,0 @@
-package com.pgms.apibooking.dto.request;
-
-import jakarta.validation.constraints.NotNull;
-
-public record SeatsGetRequest(@NotNull(message = "[공연 회차] 선택은 필수입니다.") Long eventTimeId) {
-}
