--- api/api-booking/src/main/java/com/pgms/apibooking/dto/response/PaymentCancelDetailResponse.java
@@ -1,8 +1,0 @@
-package com.pgms.apibooking.dto.response;
-
-public record PaymentCancelDetailResponse(
-	String cancelReason,
-	int cancelAmount,
-	String canceledAt
-) {
-}
