--- api/api-booking/src/main/java/com/pgms/apibooking/dto/response/PaymentCardResponse.java
@@ -1,8 +1,0 @@
-package com.pgms.apibooking.dto.response;
-
-public record PaymentCardResponse(
-	String number,
-	int installmentPlanMonths,
-	boolean isInterestFree
-) {
-}
