--- api/api-booking/src/main/java/com/pgms/apibooking/dto/response/PaymentFailResponse.java
@@ -1,8 +1,0 @@
-package com.pgms.apibooking.dto.response;
-
-public record PaymentFailResponse(
-	String errorCode,
-	String errorMessage,
-	String orderId
-) {
-}
