--- api/api-booking/src/main/java/com/pgms/apibooking/dto/response/RefundAccountResponse.java
@@ -1,8 +1,0 @@
-package com.pgms.apibooking.dto.response;
-
-public record RefundAccountResponse(
-	String bankCode,
-	String accountNumber,
-	String holderName
-) {
-}
