--- api/api-booking/src/main/java/com/pgms/apibooking/dto/response/SeatResponse.java
@@ -1,16 +1,0 @@
-package com.pgms.apibooking.dto.response;
-
-import com.pgms.coredomain.domain.event.EventSeat;
-
-public record SeatResponse(
-	String name,
-	boolean isAvailable
-) {
-
-	public static SeatResponse from(EventSeat seat) {
-		return new SeatResponse(
-			seat.getName(),
-			seat.isAvailable()
-		);
-	}
-}
