+++ api/api-booking/src/test/java/com/pgms/apibooking/fake/TossPaymentServiceFake.java
@@ -3,17 +3,17 @@ package com.pgms.apibooking.fake;
 import java.time.LocalDateTime;
 import java.util.List;
 
-import com.pgms.apibooking.dto.request.PaymentCancelRequest;
-import com.pgms.apibooking.dto.request.PaymentConfirmRequest;
-import com.pgms.apibooking.dto.response.PaymentCancelDetailResponse;
-import com.pgms.apibooking.dto.response.PaymentCancelResponse;
-import com.pgms.apibooking.dto.response.PaymentCardResponse;
-import com.pgms.apibooking.dto.response.PaymentSuccessResponse;
-import com.pgms.apibooking.dto.response.PaymentVirtualResponse;
-import com.pgms.apibooking.dto.response.RefundAccountResponse;
+import com.pgms.apibooking.domain.payment.dto.request.PaymentCancelRequest;
+import com.pgms.apibooking.domain.payment.dto.request.PaymentConfirmRequest;
+import com.pgms.apibooking.domain.payment.dto.response.PaymentCancelDetailResponse;
+import com.pgms.apibooking.domain.payment.dto.response.PaymentCancelResponse;
+import com.pgms.apibooking.domain.payment.dto.response.PaymentCardResponse;
+import com.pgms.apibooking.domain.payment.dto.response.PaymentSuccessResponse;
+import com.pgms.apibooking.domain.payment.dto.response.PaymentVirtualResponse;
+import com.pgms.apibooking.domain.payment.dto.response.RefundAccountResponse;
 import com.pgms.apibooking.exception.BookingErrorCode;
 import com.pgms.apibooking.exception.BookingException;
-import com.pgms.apibooking.service.TossPaymentService;
+import com.pgms.apibooking.domain.payment.service.TossPaymentService;
 import com.pgms.coredomain.domain.booking.Booking;
 import com.pgms.coredomain.domain.booking.Payment;
 import com.pgms.coredomain.domain.booking.PaymentStatus;
