+++ api/api-booking/src/test/java/com/pgms/apibooking/service/BookingServiceTest.java
@@ -12,11 +12,12 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Import;
 import org.springframework.transaction.annotation.Transactional;
 
+import com.pgms.apibooking.domain.booking.service.BookingService;
 import com.pgms.apibooking.config.TestConfig;
-import com.pgms.apibooking.dto.request.BookingCancelRequest;
-import com.pgms.apibooking.dto.request.BookingCreateRequest;
-import com.pgms.apibooking.dto.request.RefundAccountRequest;
-import com.pgms.apibooking.dto.response.BookingCreateResponse;
+import com.pgms.apibooking.domain.booking.dto.request.BookingCancelRequest;
+import com.pgms.apibooking.domain.booking.dto.request.BookingCreateRequest;
+import com.pgms.apibooking.domain.payment.dto.request.RefundAccountRequest;
+import com.pgms.apibooking.domain.booking.dto.response.BookingCreateResponse;
 import com.pgms.apibooking.exception.BookingErrorCode;
 import com.pgms.apibooking.exception.BookingException;
 import com.pgms.apibooking.factory.BookingFactory;
